MAINT: Add error return to all casting functionality and NpyIter
NpyIter cannot properly indicate errors right away, defering them
in some case until the actual cleanup occurs. to make that slightly
with easier, this adds a new API function with (non-standard for
NpyIter) error return as -1.
A second function is one which only checks whether an error
occurred (i.e. `NpyIter_ErrOccurred()`) which is necessary in the
(more rare) case the iterator is not deallocated promptly after
the iteration is finished.