MAINT: lib: Change handling of the expired financial functions.
In a previous commit, the expired financial functions were removed
from NumPy, and code was added to __init__.py using the module
__getattr__ to raise a customized AttributeError on any attempt to
reference the expired names in the numpy namespace.

That change broke released versions astropy, which has code that
imports the financial functions.  astropy never calls the functions,
so they never saw the deprecation warnings that have been in place
since numpy 1.18.  This means that attempting to use a released
version of astropy with numpy 1.20 will cause astropy to crash
with the custom AttributeError.

In this commit, instead of raising an exception when one of the
expired names is referenced, a warning is generated.  If the
function is *called*, an exception is raised.