BUG: Remove Void special case for "safe casting"
There were some (dubious) special cases for "safe casting"
which was only available on the C-side (np.can_cast does not
hit this code).

These special cases are almost only active for user dtypes
and seem generally incorrect and unnecessary.

Without this change, the test will fail due to promoting rational
to the void dtype (whichever that is).