DEP: Change the financial name access warning to DeprecationWarning
Most end-users will probably not care about the warning in any case,
since they will not just import the function but also use it.
We can't just remove it, due to gh-17143: Astropy currently pulls
in these functions (but doesn't use them), so that plain removal
would lead to an unusable astropy if a new NumPy is installed.
Even more annoying, due to a (faulty?) astropy pytest plugin, this
affects all pytest runs (which do not use `PYTEST_DISABLE_PLUGIN_AUTOLOAD`).

Changing it to a DeprecationWarning seems to remove the issue from
pytest runs, this may make the warning less visible in rare cases
where it should be seen, but hopefully it will still be visible enough.