BUG: Relax registration of user dtypes with `NPY_ITEM_REFCOUNT`
This also includes `NPY_ITEM_IS_POINTER`.  A previous changed
raised an error when this happened, but some downstream libraries
use it to create a custom dtype with a single object field.

It seems acceptable to create such a dtype if (and only if) that
dtype hardcodes names and fields at creation time, so this change
allows that, but otherwise keeps the error intact.

This should work fine, although some care may be required