TST: Ignore all `*` characters in the reveal tests
Mypy uses a `*` whenever an annotation or one of its parameters is based on a TypeVar.

Its added value is neglible and it unnecessarily complicates the `reveal` tests so lets just ignore them.

Note that this is done after running mypy, so it won't affect cases where `*` is used as multiplication operator.