BUG: f2py incorrectly translates dimension declarations.
In fortran functions passed to f2py, calculations that occur in the
dimensions of array declarations are interpreted as floats.  Valid
fortran requires integers.  Problem only occurs with fortran functions
not subroutines as only for the former does f2py generate fortran
wrapper code that fails to compile.

Relevant code is in numpy.f2py.crackfortran.getlincoef(), which
calculates and returns the coefficients to getarrlen() for writing to
the fortran wrapper code.

Fixes gh-8062.