MAINT: ma: Remove unused `**options` from MaskedArray `__new__` method.
The `__new__` method of `MaskedArray` has the parameter `**options`,
but it is not used.  This means a mistyped parameter is silently ignored,
e.g. if I mistype the `mask` parameter like this, I don't get an error:

    In [2]: m = np.ma.MaskedArray([1, 2, 3], maks=[0, 1, 0])

    In [3]: m
    Out[3]:
    masked_array(data=[1, 2, 3],
                 mask=False,
           fill_value=999999)

In this PR, I simply remove `**options` from the signature of
`MaskedArray.__new__`.  This affected one test.  That test was passing the
argument `shape=(0, 1, 2)` to the constructor of a subclass of `MaskedArray`,
but since any unknown keywords were silently ignored, that use of `shape` had
no effect.  I replaced it with the positional argument `[[[]], [[]]]`, which
I believe has the effect that was originally intended with the use of the
`shape` argument.