BUG: Raise promotion error if a DType was provided in array coercion (gh-17706)
The new array coercion code incorrectly promoted to "object" when a promotion was impossible and
a dtype passed in. This likely only was relevant for `dtype="V"` where different void dtypes (structured
or different length) often cannot be promoted to each other.

Previously, this worked in some cases e.g. when the array contained byte strings the correct string
was found and then cast to void. Now the void dtypes are promoted, and this fails (unless the string
lengths match).

The error for promotion is now refined in this case.