DOC: Fix a couple of reference to verbatim and vice versa
This update a coupe of references (single backticks) that actually are not to
verbatim/code (double backticks); and a couple of verbatim to reference
when they do actually exists and can be resolved in context.

I probably missed other; and stayed simple but spoted a few other
inconsistencies that I did not fix:

  - some ``...`` could actually be :math:`...` but not always clear if
  it would be better.
  - some intervals are [``...``], other are ``[...]``

I guess they could be discussed individually; it was mostly the failing
references that bothered me.