DOC: Parameter name typo axes -> axis in numpy.fft._pocketfft.
The parameter name seem to have a typo in both those case and reference
axis (and not axes), this is likely due to copy-past as some other
functions in this modules use axes (when several indices are required),
but other also use `axis` and have the correct spelling.

From review it also seem like previous phrasing is unclear so update
all similar entries to reflect the new phrasing.