Don't check if a dtype is in all_dtypes
The array API namespace is not going to do type checking against arbitrary
objects. An object that takes an array as input should assume that it will get
an array API namespace array object. Passing a NumPy array or other type of
object to any of the functions is undefined behavior, unless the type
signature allows for it.