18532: Fix the numpy Apple M1 build.
There was a build issue on numpy with M1:
AssertionError: would build wheel with unsupported tag ('cp39', 'cp39', 'macosx_11_0_universal2')

The issue was fixed in the packaging 20.5 and onwards, but that was not
getting picked by setuptools. So explicitly adding this dependency in
the toml file. The change was tested with:

```
python3 -m pip install --force-reinstall .
```