API: Formally classify `np.lib.stride_tricks` as part of the public API
With `as_strided`, and the newly introduced `sliding_window_view` function, there are currently 2 public objects that can:
a. Only be imported from a private module
b. Are publicly documented to-be imported from aforementioned module

Both observations are problematic and in need of rectification.
This commit therefore moves `np.lib.stride_tricks` to the `PUBLIC_MODULES` list.