API,DEP: Move ufunc signature parsing to the start
This may have slight affects on users, see release notes.

Mainly, we have to parse the type tuple up-front (because we need
to replace the current type resolution).  Since we _must_ de-facto
replace the current type resolution, I do not see why we should
duplicate code for the odd possibility of someone actually calling
`ufunc.type_resolver()` with a `typetup` that is not a actually a
type tuple.

This commit also deprecates `signature="l"` as meaning (normally)
the same as `dtype="l"`.