MAINT: Generalize and shorten the ufunc "trivially iterable" path
This avoids the large macros designed for specific number of
arguments and generalizes it to all ufuncs with a single output
(Instead of only 1 and 2 operand input ufuncs.)

That is not a big generalization, but the code actually shortens
and while it is slightly more complex in umath itself, it avoids
the fairly complex macros as well.

The `==` fixes a case where the fast-path was previously not
used even though it would be valid to use it. These are still
used outside the ufunc machinery.